+++ user_management/models/tests/models.py
@@ -1,7 +1,11 @@
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 
-from ..mixins import VerifyEmailMixin, AvatarMixin
+from ..mixins import AvatarMixin, BasicUserFieldsMixin, VerifyEmailMixin
 
 
 class User(AvatarMixin, VerifyEmailMixin, PermissionsMixin, AbstractBaseUser):
+    pass
+
+
+class BasicUser(BasicUserFieldsMixin, AbstractBaseUser):
     pass